requirements.txt
pdfplumber==0.11.2
pandas==2.2.2
tqdm==4.66.4
regex==2024.5.15
rapidfuzz==3.9.7
python-dotenv
google-generativeai
